Moving FastLogger registration to WebHost
