+++ src/WebJobs.Script.NuGet/WebJobs.Script.nuspec
@@ -16,7 +16,6 @@
     <dependencies>
       <dependency id="Microsoft.Azure.WebJobs" version="1.2.0-alpha-10306" />
       <dependency id="Microsoft.Azure.WebJobs.ServiceBus" version="1.2.0-alpha-10306" />
-      <dependency id="Microsoft.Azure.WebJobs.Logging" version="1.2.0-alpha-10306" />
       <dependency id="Microsoft.Azure.WebJobs.Extensions" version="1.1.0-alpha-10287" />
       <dependency id="Microsoft.Azure.WebJobs.Extensions.SendGrid" version="1.1.0-alpha-10287" />
       <dependency id="Microsoft.Azure.WebJobs.Extensions.ApiHub" version="1.0.0-alpha-10287" />
