+++ src/WebJobs.Script.WebHost/Diagnostics/FastLogger.cs
@@ -4,12 +4,14 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Host.Loggers;
 using Microsoft.Azure.WebJobs.Logging;
+using Microsoft.Azure.WebJobs.Script;
 using Microsoft.WindowsAzure.Storage;
 using Newtonsoft.Json;
 
-namespace Microsoft.Azure.WebJobs.Script.Diagnostics
+namespace WebJobs.Script.WebHost.Diagnostics
 {
     // Adapter for capturing SDK events and logging them to tables.
     internal class FastLogger : IAsyncCollector<FunctionInstanceLogEntry>
