+++ src/WebJobs.Script.WebHost/packages.config
@@ -35,6 +35,7 @@
   <package id="Microsoft.Azure.WebJobs.Extensions.MobileApps" version="1.0.0-alpha-10287" targetFramework="net46" />
   <package id="Microsoft.Azure.WebJobs.Extensions.NotificationHubs" version="1.0.0-alpha-10287" targetFramework="net46" />
   <package id="Microsoft.Azure.WebJobs.Extensions.SendGrid" version="1.1.0-alpha-10287" targetFramework="net46" />
+  <package id="Microsoft.Azure.WebJobs.Logging" version="1.2.0-alpha-10306" targetFramework="net46" />
   <package id="Microsoft.Azure.WebJobs.ServiceBus" version="1.2.0-alpha-10306" targetFramework="net46" />
   <package id="Microsoft.Bcl" version="1.1.10" targetFramework="net46" />
   <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="net46" />
